lab-02/requirements.txt
# RAG pipeline dependencies
numpy>=1.24.0
faiss-cpu>=1.7.0
python-dotenv>=1.0.0
google-genai>=1.0.0

# API
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
